% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3_calling.R
\name{abs_calling}
\alias{abs_calling}
\title{absCopyNumber calling with user specified parameters}
\usage{
abs_calling(absCopyNumber, samples = NULL, verbose = FALSE)
}
\arguments{
\item{absCopyNumber}{a \code{absCopyNumber} object generate from \code{abs_prepare} function.}

\item{samples}{character vector to subset samples for calling, default is \code{NULL}.}

\item{verbose}{if \code{TRUE}, print extra information.}
}
\value{
a \code{absCopyNumber} object with calling result.
}
\description{
This is the third step for absCopyNumber pipeline, absCopyNumber compute
solutions of purity and ploidy, then generate absolute copynumber and (optional) multiplicity
for top 1 solution.
}
\examples{
#' file_cn = system.file("extdata/example.cn.txt.gz", package = "absCopyNumber")
file_snv = system.file("extdata/example.snv.txt.gz", package = "absCopyNumber")

\donttest{
res1 =  abs_initialize(seg = file_cn, snv = file_snv, verbose = T)
res1 =  abs_prepare(res1)
res1 =  abs_calling(res1, verbose = TRUE)
}

}
\seealso{
\code{\link[absCopyNumber]{absCopyNumber}}, \code{\link[absCopyNumber]{abs_initialize}}, , \code{\link[absCopyNumber]{abs_prepare}}
}
\author{
Shixiang Wang <w_shixiang@163.com>
}
