% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3_calling.R
\name{abs_obtain}
\alias{abs_obtain}
\title{Obtain absolute copy number and optinal multiplicity (if SNV data provided)}
\usage{
abs_obtain(object, rank = 1, samples = NULL, onlyCN = FALSE,
  verbose = FALSE)
}
\arguments{
\item{object}{a \code{absCopyNumber} object after absCopyNumber calling}

\item{rank}{integer, rank value of solution, i.e. \code{estimation} slot.}

\item{samples}{character vector, use if you wanna just get solution for specified samples.}

\item{onlyCN}{if \code{TRUE}, just get absolute copy number result (a \code{data.table}), otherwise
return a list include both absolute copy number and snv data with multiplicity.}

\item{verbose}{if \code{TRUE}, print extra information.}
}
\value{
a \code{list} or a \code{data.table}, details see \code{onlyCN} option.
}
\description{
According to the user specified \code{rank} value, this function retrieve corresponding
purity and ploidy solution pair, then compute absolute copy number, optional muliplicity.
}
\author{
Shixiang Wang <w_shixiang@163.com>
}
