% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_prepare.R
\name{abs_prepare}
\alias{abs_prepare}
\title{Prepare parameters for absCopyNumber calling}
\usage{
abs_prepare(absCopyNumber, alpha.min = 0.2, alpha.max = 1,
  tau.min = 0.5, tau.max = 8, copyratio.min = -3,
  copyratio.max = 3, qmax = 10, lamda = 0.5,
  method = c("Grid Search", "Bayesian Optimization"),
  min.sol.freq = 0.05, snv.type = c("somatic", "germline"),
  force = FALSE)
}
\arguments{
\item{absCopyNumber}{a \code{absCopyNumber} object generated by \code{abs_initialize} function.}

\item{alpha.min}{the minimum allowed
value for tumor purity. Default is 0.20. If you do have the pathologist
estimate, set it as the lower bound of the pathologist estimate is usually
preferred.}

\item{alpha.max}{the maximum allowed
value for tumor purity. Default is 1.0. If you do have the pathologist
estimate, set it as the upper bound of the pathologist estimate is usually
preferred.}

\item{tau.min}{the minimum allowed value for tumor ploidy, default is 0.5.}

\item{tau.max}{the maximum allowed value for tumor ploidy, default is 8.}

\item{copyratio.min}{the minimum value of copy number ratio to assign weight in model, default is -3.
Of note, if platform is 'WES' or 'WGS', to filter noise in NGS, \code{copyratio.min}
 will be changed to to 1/copyratio.max, it means this argument will not be used. If
 you are sure what you are doing, you can turn argument \code{force} to \code{TRUE} to
 make sure \code{copyration.min} use what you are input.}

\item{copyratio.max}{the maximum value of copy number ratio to assign weight in model, default is 3.}

\item{qmax}{maximum allowed absolute copy number for any segments, default is 10.}

\item{lamda}{The relative weight of the
segment copy ratio data over the SNV data. Must be a value in (0.0,1.0]. Only used when SNV
file is provided. The default value is 0.5,
which give equal weights to copy-number-based ratio estimator and SNV-frequency-based estimator. If
\code{lamda} is 1, only use copy-number-based ratio model to minimized objective function.}

\item{method}{method for object function optimization, should be one of "Grid Search", "Bayesian Optimization".}

\item{min.sol.freq}{A solution
should appear at least this proportion to be kept. Singleton solutions are
usually not trustable. By default (min.sol.freq=0.05), the program will only
retain solutions that cover at least 5 percent of the search space.}

\item{snv.type}{one of "somatic", "germline", if no SNV data, just let is be "somatic"}

\item{force}{if \code{TRUE}, force \code{copyratio.min} as exactly user's input.}
}
\value{
a \code{absCopyNumber} object with parameters for model construction.
}
\description{
This is the second step for absCopyNumber pipeline, user provide
parameters for model construction.
}
\examples{
#' file_cn = system.file("extdata/example.cn.txt.gz", package = "absCopyNumber")
file_snv = system.file("extdata/example.snv.txt.gz", package = "absCopyNumber")

\donttest{
res1 =  abs_initialize(seg = file_cn, snv = file_snv, verbose = T)
res1 =  abs_prepare(res1)
}

}
\seealso{
\code{\link[absCopyNumber]{absCopyNumber}}, \code{\link[absCopyNumber]{abs_initialize}}, , \code{\link[absCopyNumber]{abs_calling}}
}
\author{
Shixiang Wang <w_shixiang@163.com>
}
