% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_prepare.R
\name{abs_subset}
\alias{abs_subset}
\title{subset a absCopyNumber object by samples}
\usage{
abs_subset(object, samples = NULL)
}
\arguments{
\item{object}{a \code{absCopyNumber} object}

\item{samples}{a character vector to specify samples.}
}
\value{
a \code{absCopyNumber} object
}
\description{
subset a absCopyNumber object by samples
}
\author{
Shixiang Wang <w_shixiang@163.com>
}
