% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{abs_supportfiles}
\alias{abs_supportfiles}
\title{Print information about input of absCopyNumber and extra message}
\usage{
abs_supportfiles()
}
\value{
Nothing
}
\description{
Print information about input of absCopyNumber and extra message
}
\examples{
abs_supportfiles()
}
