% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4class.R
\docType{class}
\name{absCopyNumber-class}
\alias{absCopyNumber-class}
\alias{absCopyNumber}
\title{Class absCopyNumber}
\description{
S4 class for storing absCopyNumber input and results
}
\section{Slots}{

\describe{
\item{\code{data}}{a \code{data.table} of Segmentation file containing all filtered segments}

\item{\code{SNV}}{a \code{data.table} containing input single nucletide variants.}

\item{\code{params}}{a list containing parameters control data filter and model construction etc.}

\item{\code{estimation}}{table containing estimated results ranked by their fitting errors}

\item{\code{TopResult}}{a list containing result with minimal fitting errors, include 2 \code{data.table}:
absCN - table containing the absolute copy number estimates of tumor cells based on purity and ploidy pair;
absSNV - table containing the absolute multiplicity estimates of SNVs based on purity and ploidy pair.}

\item{\code{origin}}{a data.table contain original segmentation data}
}}

