% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absCopyNumber.R
\name{get_absCopyNumber}
\alias{get_absCopyNumber}
\title{Convert copy ratios into abosolute copy numbers when given the purity and ploidy estimates}
\usage{
get_absCopyNumber(seg.data, alpha, tau, qmax = 7)
}
\arguments{
\item{seg.data}{a data frame with five
columns: "chrom", "loc.start", "loc.end", "eff.seg.len", "normalized.ratio".}

\item{alpha}{tumor purity estimate}

\item{tau}{tumor ploidy estimate}

\item{qmax}{maximum allowed absolute copy
number for any segments}
}
\value{
a \code{data.frame}
columns: rhat (expected copy ratio) and CN (absolute copy number)
}
\description{
This function converts copy ratios into abosolute copy numbers when the tumor
purity and ploidy are provided by the user. The tumor purity and ploidy
parameter should be taken from the result after running the "run.absCNSeq"
function.
}
\author{
Lei Bao
}
