% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absCopyNumber.R
\name{plot_absCopyNumber}
\alias{plot_absCopyNumber}
\title{Plot estimated integer copy numbers for each chromosome.}
\usage{
plot_absCopyNumber(seg.data, chromnum = 1, rawdata = NULL,
  chromvar = "chrom", loc.start = "chr_start", loc.end = "chr_stop",
  normalized.ratio = "log2_ratio")
}
\arguments{
\item{seg.data}{A result data frame
generated by the compute.absCN function. It has five mandatory columns:
"chrom", "loc.start", "loc.end", "r", "CN".}

\item{chromnum}{which chromosome to be
plotted. Must be a number from 1 to 22}

\item{rawdata}{(optional): raw copy
ratio data before segmentation. This data frame has four mandatory columns:
chromosome number, start position, end position, and normalized copy ratio
of tumor vs. normal sample on log2 scale (log2_ratio). The column names are
provided by the user. Default column names are taken from the VARSCAN
output.}

\item{chromvar}{The variable name of
the chromosome number column for the optional raw data.}

\item{loc.start}{The variable name of
the start position column for the optional raw data.}

\item{loc.end}{The variable name of the
end position column for the optional raw data.}

\item{normalized.ratio}{The
variable name of the log2_ratio column for the optional raw data.}
}
\description{
plot estimated integer copy numbers from absCNseq for each chromosome.
}
\author{
Lei Bao
}
