% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absCopyNumber.R
\name{run_absCopyNumber}
\alias{run_absCopyNumber}
\title{Main function to run AbsoluteCopyNumber}
\usage{
run_absCopyNumber(seg.fn, snv.fn = NULL, res.dir, smp.name,
  seq.type = c("WES", "WGS"), alpha.min = 0.2, alpha.max = 1,
  tau.min = 1.5, tau.max = 5, min.sol.freq = 0, min.seg.len = 0,
  qmax = 7, lamda = 0.5, verbose = FALSE)
}
\arguments{
\item{seg.fn}{The name of the file
containing the segmentation data. An example file called "example.cn.txt"
can be found under the "example" folder within the tarball.}

\item{snv.fn}{The name of the file
containing a set of somatic single nucleotide variants (SNVs). An example
file called "example.snv.txt" can be found under the "example" folder within
the tarball.}

\item{res.dir}{The output directory}

\item{smp.name}{Sample name}

\item{seq.type}{Either "WES" (whole exome sequencing) or "WGS" (whole genome sequencing)}

\item{alpha.min}{The minimum allowed
value for tumor purity. Default is 0.20. If you do have the pathologist
estimate, set it as the lower bound of the pathologist estimate is usually
preferred.}

\item{alpha.max}{The maximum allowed
value for tumor purity. Default is 1.0. If you do have the pathologist
estimate, set it as the upper bound of the pathologist estimate is usually
preferred.}

\item{tau.min}{The minimum allowed value
for tumor ploidy}

\item{tau.max}{The maximum allowed value
for tumor ploidy}

\item{min.sol.freq}{A solution
should appear at least this many times to be kept. Singleton solutions are
usually not trustable. By default (min.sol.freq=0), the program will only
retain solutions that cover at least 1 percent of the search space.}

\item{min.seg.len}{The minimum
length of a segment to be included in computation. The default value is 200
bp for WES and 3000 bp for WGS.}

\item{qmax}{Maximum allowed absolute copy
number for any segments.}

\item{lamda}{The relative weight of the
segment copy ratio data over the SNV data. Must be a value in(0.0,1.0].}

\item{verbose}{if \code{True}, print detail information}
}
\value{
\item{searchRes }{a data frame giving the solution set (purity and
ploidy pairs) ranked by their fitting errors }
\item{absCN }{a data frame
giving the absolute copy number estimates of tumor cells for the top first
solution (purity and ploidy pair)}
\item{absSNV }{a data frame giving the
absolute multiplicity estimates of SNVs for the top first solution (purity
and ploidy pair) }
\item{orig.seg.dat }{original copy ratio data read from
the input segmentation file }
\item{seg.dat }{filtered copy ratio data }
\item{orig.snv.dat }{original SNV data read from the input SNV file }
\item{snv.dat }{filtered SNV data }
}
\description{
This wrapper function accepts data files and user specified parameters and runs
the absCNseq algorithm from paper "AbsCN-seq: a statistical method to estimate
tumor purity, ploidy and absolute copy numbers from next-generation sequencing data."
}
\details{
Please refer to the "example" folder to see the format of the input segmentation file or
SNV file.
}
\references{
AbsCN-seq: a statistical method to estimate tumor purity,
ploidy and absolute copy numbers from next-generation sequencing data.
}
\author{
Lei Bao
}
