% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absCopyNumber.R
\name{run_fromDF}
\alias{run_fromDF}
\title{Run absCopyNumber from data.frame}
\usage{
run_fromDF(seg.df, snv.df = NULL, platform = c("WES", "WGS",
  "MicroArray"), alpha.min = 0.2, alpha.max = 1, tau.min = 0.5,
  tau.max = 8, min.sol.freq = 0, min.seg.len = NULL, qmax = 7,
  lamda = 0.5, verbose = FALSE)
}
\arguments{
\item{seg.df}{The data.frame
containing the segmentation data. An example file called "example.cn.txt.gz"
can be found under the "inst/extdata" folder within the package.}

\item{snv.df}{The data.frame
containing a set of somatic single nucleotide variants (SNVs). An example
file called "example.snv.txt.gz" can be found under the "inst/extdata" folder within
the package.}

\item{platform}{"WES" for whole exome sequencing, "WGS" for whole genome sequencing and
"MicroArray" for SNP or other gene chips can detect CNV.}

\item{alpha.min}{The minimum allowed
value for tumor purity. Default is 0.20. If you do have the pathologist
estimate, set it as the lower bound of the pathologist estimate is usually
preferred.}

\item{alpha.max}{The maximum allowed
value for tumor purity. Default is 1.0. If you do have the pathologist
estimate, set it as the upper bound of the pathologist estimate is usually
preferred.}

\item{tau.min}{The minimum allowed value for tumor ploidy, default is 0.5.}

\item{tau.max}{The maximum allowed value for tumor ploidy, default is 8.}

\item{min.sol.freq}{A solution
should appear at least this many times to be kept. Singleton solutions are
usually not trustable. By default (min.sol.freq=0), the program will only
retain solutions that cover at least 1 percent of the search space.}

\item{min.seg.len}{The minimum
length of a segment to be included in computation. The default value is 200
bp for WES, 3000 bp for WGS and 0 bp for MicroArray.}

\item{qmax}{Maximum allowed absolute copy number for any segments, default is 7.}

\item{lamda}{The relative weight of the
segment copy ratio data over the SNV data. Must be a value in(0.0,1.0]. Only used when SNV
file is provided. The default value is 0.5,
which give equal weights to copy-number-based ratio estimator and SNV-frequency-based estimator. If
\code{lamda} is 1, only use copy-number-based ratio model to minimized objective function.}

\item{verbose}{if \code{True}, print extra output}
}
\value{
\item{searchRes }{a data frame giving the solution set (purity and
ploidy pairs) ranked by their fitting errors }
\item{absCN }{a data frame
giving the absolute copy number estimates of tumor cells for the top first
solution (purity and ploidy pair)}
\item{absSNV }{a data frame giving the
absolute multiplicity estimates of SNVs for the top first solution (purity
and ploidy pair) }
\item{orig.seg.dat }{original copy ratio data read from
the input segmentation file }
\item{seg.dat }{filtered copy ratio data }
\item{orig.snv.dat }{original SNV data read from the input SNV file }
\item{snv.dat }{filtered SNV data }
}
\description{
This wrapper function accepts data.frame and user specified parameters and runs
the absCNseq algorithm from paper "AbsCN-seq: a statistical method to estimate
tumor purity, ploidy and absolute copy numbers from next-generation sequencing data."
}
\references{
AbsCN-seq: a statistical method to estimate tumor purity,
ploidy and absolute copy numbers from next-generation sequencing data.
}
\author{
Lei Bao, Shixiang Wang
}
