% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1_initialize.R
\name{abs_initialize}
\alias{abs_initialize}
\title{Initialize absCopyNumber object from file or data.frame}
\usage{
abs_initialize(seg, min.seg.len = NULL, snv = NULL, isMaf = FALSE,
  sample_seg = NULL, sample_snv = NULL, ratio.min = -3,
  ratio.max = 3, platform = c("WES", "WGS", "MicroArray"),
  verbose = FALSE)
}
\description{
This is the first step for user to use a flexible way
}
