% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4class.R
\docType{class}
\name{absCopyNumber-class}
\alias{absCopyNumber-class}
\alias{absCopyNumber}
\title{Class absCopyNumber}
\description{
S4 class for storing input and results
}
\section{Slots}{

\describe{
\item{\code{data}}{data.table of Segmentation file containing all filtered segments}

\item{\code{snv.data}}{table containing all filtered single nucletide variants}

\item{\code{params}}{a list containing parameters control data filter and model construction etc.}

\item{\code{origin}}{a list contain original segmentation and snv data}

\item{\code{summary}}{table with basic input data stats}

\item{\code{result}}{table containing estimated results ranked by their fitting errors}

\item{\code{absCN}}{table containing the absolute copy number estimates of tumor cells for the top first solution (purity and ploidy pair)}

\item{\code{absSNV}}{table containing the absolute multiplicity estimates of SNVs for the top first solution (purity and ploidy pair)}
}}

\seealso{
\code{\link{abs_generator}}
}
